requirements.txt
torch>=2.0
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12
allensdk>=2.16
jupyter>=1.0
ipykernel>=6.0
pyyaml>=6.0
tqdm>=4.65
